ChemOrch/generation/requirements.txt
anthropic==0.51.0
openai==1.79.0
pandas==2.2.3
pubchempy==1.0.4
python-dotenv==1.1.0
PyYAML==6.0.2
PyYAML==6.0.2
tiktoken==0.9.0
torch==2.6.0
